Move Field enum & preprocessors to separate helper modules (#11)
* [change] move Field and preprocessors to separate helper modules

* [change] tests
